requirements.txt
altgraph==0.17.4
colorama==0.4.6
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
qrcode==7.4.2
setuptools==69.5.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
